refactor: extract user repository
