+++ src/backend/Cargo.toml
@@ -7,6 +7,14 @@ edition = "2021"
 
 [dependencies]
 
+[dependencies.uuid]
+version = "1.1.2"
+features = [
+    "v4",                # Lets you generate random UUIDs
+    "fast-rng",          # Use a faster (but still sufficiently random) RNG
+    "macro-diagnostics", # Enable better diagnostics for compile-time UUIDs
+]
+
 [package.metadata.precommit]
 fmt = "cargo fmt -- --write-mode diff 2>&1"
 test = "cargo test 2>&1"
