+++ src/backend/src/infra.rs
@@ -1,0 +1,1 @@
+pub mod user_repository;
