+++ src/backend/src/infra/user_repository.rs
@@ -1,0 +1,21 @@
+use crate::domain::user::User;
+use crate::services::login::UserRepository;
+
+pub struct UserRepositoryInMemory {
+    users: Vec<User>,
+}
+
+impl UserRepository for UserRepositoryInMemory {
+    fn find_user_by_email(&self, email: &str) -> Option<&User> {
+        self.users
+            .iter()
+            .find(|user| user.email == email.to_string())
+    }
+
+    fn new() -> Self {
+        let user = User::new("test@test.com", "1234566", "test");
+
+        let users: Vec<User> = vec![user];
+        Self { users }
+    }
+}
