+++ src/backend/src/main.rs
@@ -1,5 +1,6 @@
 mod domain;
 mod services;
+mod infra;
 
 pub use crate::services::login;
