+++ src/backend/src/services/login.rs
@@ -1,7 +1,8 @@
 pub use crate::domain::{token::Token, user::User};
+pub use crate::infra::user_repository::UserRepositoryInMemory;
 
-trait UserRepository {
-    fn findUserByEmail(&self, email: &str) -> Option<&User>;
+pub trait UserRepository {
+    fn find_user_by_email(&self, email: &str) -> Option<&User>;
     fn new() -> Self;
 }
 
@@ -16,7 +17,7 @@ fn handle(
     generate_token: fn(id: &str) -> String,
     repo: impl UserRepository,
 ) -> Result<Token, LoginError> {
-    let user: Option<&User> = repo.findUserByEmail(email);
+    let user: Option<&User> = repo.find_user_by_email(email);
 
     if user.is_none() {
         return Err(LoginError::InvalidCrendetias);
@@ -27,34 +28,14 @@ fn handle(
     })
 }
 
-struct UserRepositoryInMemory {
-    users: Vec<User>,
-}
-
-impl UserRepository for UserRepositoryInMemory {
-    fn findUserByEmail(&self, email: &str) -> Option<&User> {
-        self.users
-            .iter()
-            .find(|user| user.email == email.to_string())
-    }
-
-    fn new() -> Self {
-        let user = User {
-            id: String::from("a64f634a-5b2c-4c0a-9fc3-cd448fedc5b0"),
-            email: String::from("test@test.com"),
-            password: String::from("1234566"),
-            name: String::from("test"),
-        };
-
-        let users: Vec<User> = vec![user];
-        Self { users }
-    }
-}
-
 #[cfg(test)]
 mod tests_services {
 
     use super::*;
+
+    fn generate_token(_id: &str) -> String {
+        String::from("test_token")
+    }
 
     #[test]
     fn should_create_a_token_when_give_a_valid_credentials() {
@@ -65,7 +46,6 @@ mod tests_services {
         let email = "test@test.com";
         let password = "1234566";
 
-        let generate_token = |id: &str| String::from("test_token");
         let repo = UserRepositoryInMemory::new();
 
         let token = handle(email, password, generate_token, repo).expect("token should be valid");
@@ -77,8 +57,6 @@ mod tests_services {
     fn should_return_a_error_if_user_not_exits() {
         let email = "notexist@test.com";
         let password = "1234566";
-
-        let generate_token = |id: &str| String::from("test_token");
 
         let repo = UserRepositoryInMemory::new();
